Extract neutral workflow progress facts
