+++ ui/application/__init__.py
@@ -109,8 +109,13 @@ from .local_first_control_visibility import (
 )
 from .local_first_filter_summary import build_local_first_filter_description
 from .local_first_progress_facts import (
+    LocalFirstProgressFacts,
     current_local_first_last_sync_date,
     runtime_state_with_local_first_output_path,
+)
+from .workflow_progress_facts import (
+    WorkflowProgressFacts,
+    build_workflow_progress_facts_from_runtime_state,
 )
 from .dock_workflow_sections import (
     CURRENT_DOCK_SECTIONS,
@@ -215,6 +220,7 @@ __all__ = [
     "LocalFirstDockPageDefinition",
     "LocalFirstDockPageState",
     "LocalFirstParitySurface",
+    "LocalFirstProgressFacts",
     "LocalFirstWidgetMove",
     "WIZARD_WORKFLOW_STEPS",
     "WIZARD_STEP_COUNT",
@@ -232,6 +238,7 @@ __all__ = [
     "VisualWorkflowActionInputs",
     "VisualWorkflowSettingsSnapshot",
     "WizardProgressFacts",
+    "WorkflowProgressFacts",
     "WizardSettingsSnapshot",
     "build_advanced_fetch_visibility_update",
     "apply_local_first_visibility_update",
@@ -262,6 +269,7 @@ __all__ = [
     "build_startup_wizard_progress_facts",
     "current_local_first_last_sync_date",
     "build_wizard_filter_description",
+    "build_workflow_progress_facts_from_runtime_state",
     "build_wizard_progress_facts_from_runtime_state",
     "build_wizard_progress_from_facts",
     "build_wizard_progress_from_facts_and_settings",
